chore: make structure neat
