+++ src/App.tsx
@@ -1,6 +1,5 @@
 import { ChakraProvider } from '@chakra-ui/react'
 import { Main } from './controllers/Main'
-import './App.css'
 
 export function App() {
   return (
