+++ src/components/Form/index.tsx
@@ -1,71 +1,15 @@
 import React from 'react'
 
-import {
-  Checkbox,
-  Input,
-  NumberInput,
-  NumberInputField,
-  NumberInputStepper,
-  NumberIncrementStepper,
-  NumberDecrementStepper,
-  Radio,
-  Textarea,
-  Box,
-  RadioGroup,
-  Stack,
-  Button,
-  Text,
-} from '@chakra-ui/react'
-import { capitalizeFirstLetter } from '../../utils/capitalize-first-letter'
+import { Box, Stack, Button, Text } from '@chakra-ui/react'
 
 import { formAdapter } from '../../utils/form-adapter'
-import { FormItem, FieldTypes } from '../../interfaces'
+import { FormField } from '../../interfaces'
 
-import { FormItemWrapper } from '../FormItemWrapper'
+import { FormElementContainer } from '../ui/FormElementContainer'
+import { FormElement } from '../ui/FormElement'
 
 interface IProps {
   jsonInput: string
-}
-
-const renderFormElement = (item: FormItem) => {
-  switch (item.type) {
-    case FieldTypes.CHECKBOX:
-      return (
-        <Checkbox id={item.label} size="lg">
-          {capitalizeFirstLetter(item.label)}
-        </Checkbox>
-      )
-    case FieldTypes.TEXT:
-      return <Input id={item.label} />
-    case FieldTypes.NUMBER:
-      return (
-        <NumberInput id={item.label}>
-          <NumberInputField />
-          <NumberInputStepper>
-            <NumberIncrementStepper />
-            <NumberDecrementStepper />
-          </NumberInputStepper>
-        </NumberInput>
-      )
-    case FieldTypes.DATE:
-      return <input type="date" id={item.label} />
-    case FieldTypes.RADIO:
-      return (
-        <RadioGroup>
-          <Stack spacing={4} direction="row">
-            {item.options?.map((radio: string) => (
-              <Radio value={radio} key={radio}>
-                {radio}
-              </Radio>
-            ))}
-          </Stack>
-        </RadioGroup>
-      )
-    case FieldTypes.TEXTAREA:
-      return <Textarea id={item.label} />
-    default:
-      return <Input id={item.label} />
-  }
 }
 
 // не забыть плейсхолдеры?
@@ -85,8 +29,10 @@ export const Form: React.FC<IProps> = ({ jsonInput }) => {
         </Text>
       )}
       <form onSubmit={onSubmit}>
-        {formFields.map((item: FormItem) => (
-          <FormItemWrapper item={item}>{renderFormElement(item)}</FormItemWrapper>
+        {formFields.map((field: FormField) => (
+          <FormElementContainer field={field}>
+            <FormElement field={field} />
+          </FormElementContainer>
         ))}
         <Stack direction="row" mt="10" justifyContent="space-between">
           {buttons.map((button: string) => (
