+++ src/controllers/Main/hooks/useEditor.ts
@@ -1,5 +1,5 @@
 import React, { useState } from 'react'
-import { prettyPrint } from '../utils'
+import { prettyPrint } from '../../../utils/pretty-print-json'
 
 import { defaultInputValue } from './defaults'
