+++ src/controllers/Main/index.tsx
@@ -3,8 +3,8 @@ import { Editor } from '../../components/Editor'
 import { Tabs } from '../../components/Tabs'
 import { Form } from '../../components/Form'
 
-import { useEditor } from '../../hooks/useEditor'
-import { useTabs } from '../../hooks/useTabs'
+import { useEditor } from './hooks/useEditor'
+import { useTabs } from './hooks/useTabs'
 import { TabTypes } from '../../interfaces'
 
 export const Main = () => {
