--- src/utils/index.ts
@@ -1,6 +1,0 @@
-export function prettyPrint(value: string) {
-  const obj = JSON.parse(value)
-  const pretty = JSON.stringify(obj, undefined, 4)
-
-  return pretty
-}
